--- src/main/java/com/lxp/lecture/dto/LectureCreateResponse.java
@@ -1,6 +1,0 @@
-package com.lxp.lecture.dto;
-
-public record LectureCreateResponse(
-        Long id
-) {
-}
