+++ src/main/java/com/lxp/lecture/presentation/controller/dto/response/LectureCreateResponse.java
@@ -1,0 +1,6 @@
+package com.lxp.lecture.presentation.controller.dto.response;
+
+public record LectureCreateResponse(
+        Long id
+) {
+}
