+++ src/test/java/com/lxp/lecture/service/LectureSaveTest.java
@@ -12,10 +12,10 @@ import static org.mockito.Mockito.when;
 import com.lxp.course.dao.CourseDao;
 import com.lxp.course.exception.CourseNotFoundException;
 import com.lxp.lecture.dao.LectureDao;
-import com.lxp.lecture.dto.LectureCreateRequest;
-import com.lxp.lecture.dto.LectureCreateResponse;
 import com.lxp.lecture.exception.LectureNotSavedException;
 import com.lxp.lecture.model.Lecture;
+import com.lxp.lecture.presentation.controller.dto.request.LectureCreateRequest;
+import com.lxp.lecture.presentation.controller.dto.response.LectureCreateResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
